Update index.html
Your HTML code looks well-structured for a landing page. However, I noticed a minor issue with the closing </ul> tag for the "dynamic-text" unordered list. It seems to be missing, and it's important to close the list properly

Find your Space button, it might be a good idea to use a type= button attribute within the button tag to explicitly define it as a button